requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==1.2.2
black==21.12b0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
contourpy==1.0.5
cryptography==40.0.1
cycler==0.11.0
cytoolz==0.11.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-allauth==0.54.0
django-appconf==1.0.5
django-compressor==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
entrypoints==0.4
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.10.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
flatbuffers==2.0.7
fonttools==4.37.2
frozenlist==1.2.0
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
hexbytes==0.2.2
idna==3.4
ipfshttpclient==0.8.0a2
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
libretranslatepy==2.1.1
lru-dict==1.1.7
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
nltk==3.8.1
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2023.3.20
platformdirs==3.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.1
psutil==5.9.3
pure-eval==0.2.2
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytube==12.1.0
pytz==2023.3
pywin32==303
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
rcssmin==1.1.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
rlp==2.0.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.3
stack-data==0.5.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==1.2.3
toolz==0.11.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
translate==3.6.1
typing_extensions==4.0.1
tzdata==2023.2
urllib3==1.26.15
varint==1.0.2
virtualenv==20.21.0
virtualenv-clone==0.5.7
watchdog==2.1.9
wcwidth==0.2.5
web3==5.25.0
webencodings==0.5.1
websocket-client==1.4.1
websockets==9.1
Werkzeug==2.2.2
whitenoise==6.4.0
widgetsnbextension==4.0.3
word2number==1.1
wrapt==1.14.1
yarl==1.7.2
